Все еще функционал находится в одном микросервисе.  Использованы  REST контроллеры.  Добавил мапперы для DTO ModelMapper  Добавил обработки ошибок @ExceptionHandler.  Добавил jwt аутентификациЮ.  Убрал @Transactional там, где это не требуется.

+++ library/library/src/main/java/skr/library/services/BookService.java
@@ -1,11 +1,12 @@
 package skr.library.services;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import skr.library.models.Book;
 import skr.library.repositories.BookRepository;
+import skr.library.util.BookAlreadyExistException;
+import skr.library.util.BookNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -32,40 +33,47 @@ public class BookService {
     // Получение книги по ID
     @Transactional(readOnly = true)
     public Book getBookById(int id) {
-        Optional<Book> foundBook = bookRepository.findById(id);
-        return foundBook.orElse(null);
+        Optional<Book> book = bookRepository.findById(id);
+        return book.orElseThrow(BookNotFoundException::new);
     }
 
     // Получение книги по ISBN
     @Transactional(readOnly = true)
     public Book getBookByIsbn(String isbn) {
-        Optional<Book> foundBook = bookRepository.findByIsbn(isbn);
-        return foundBook.orElse(null);
+        Optional<Book> book = bookRepository.findByIsbn(isbn);
+        return book.orElseThrow(BookNotFoundException::new);
     }
 
-    // Добавление новой книги(свободной)
-    @Transactional
-    public void addBook(Book book) {
+    // Добавление новой книги (свободной)
+    public Book addBook(Book book) {
+        Optional<Book> existingBook = bookRepository.findByIsbn(book.getIsbn());
+        if (existingBook.isPresent()) {
+            throw new BookAlreadyExistException();
+        }
         Book savedBook = bookRepository.save(book);
         libraryService.addBookAsAvailable(savedBook.getId());
+        return savedBook;
     }
 
-    // Изменение существующей книги
-    public void updateBook(int id, Book updatedBook) {
-        Optional<Book> optionalBook = bookRepository.findById(id);
-        if (optionalBook.isPresent()) {
-            Book book = optionalBook.get();
-            book.setIsbn(updatedBook.getIsbn());
-            book.setTitle(updatedBook.getTitle());
-            book.setGenre(updatedBook.getGenre());
-            book.setDescription(updatedBook.getDescription());
-            book.setAuthor(updatedBook.getAuthor());
-            bookRepository.save(book);
-        }
+    //Обновление информации о книше по id
+    public Book updateBook(int id, Book updatedBook) {
+        return bookRepository.findById(id)
+                .map(book -> {
+                    book.setIsbn(updatedBook.getIsbn());
+                    book.setTitle(updatedBook.getTitle());
+                    book.setGenre(updatedBook.getGenre());
+                    book.setDescription(updatedBook.getDescription());
+                    book.setAuthor(updatedBook.getAuthor());
+                    return bookRepository.save(book);
+                })
+                .orElseThrow(BookNotFoundException::new);
     }
+
     // Удаление книги
-    @Transactional
     public void deleteBook(int id) {
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException();
+        }
         bookRepository.deleteById(id);
         libraryService.deleteLibrary(id);
     }
